requirements.txt
playwright>=1.50.0
python-docx>=1.1.0
pyinstaller>=6.0.0
